// frontend/src/app/models/auth.models.ts
export interface User {
    id: string;
    name: string;
    email: string;
    profileId: string;
}

export interface AuthResponse {
    user: User;
    token: string;
}

export interface LoginRequest {
    email: string;
    password: string;
}

export interface RegisterRequest {
    name: string;
    email: string;
    password: string;
    profileId: string;
}